Move number settings to NumberSettings
